[CHORE#1606] onboarding , navigation 파일들 디렉토리 경로 변경
