+++ feature/sopletter/src/main/java/org/sopt/official/feature/sopletter/onboarding/SopletterOnboardingScreen.kt
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.sopletter.onboarding
+package org.sopt.official.feature.sopletter.onboarding
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -36,8 +36,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.sopt.official.designsystem.SoptTheme
-import org.sopt.official.sopletter.component.SopletterButton
-import org.sopt.official.sopletter.onboarding.component.SopletterOnboardingInfoHolder
+import org.sopt.official.feature.sopletter.common.component.SopletterButton
+import org.sopt.official.feature.sopletter.onboarding.component.SopletterOnboardingInfoHolder
 
 @Composable
 fun SopletterOnboardingRoute(
