+++ feature/sopletter/src/main/java/org/sopt/official/feature/sopletter/onboarding/component/SopletterOnboardingInfoHolder.kt
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.sopletter.onboarding.component
+package org.sopt.official.feature.sopletter.onboarding.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
