+++ feature/sopletter/src/main/java/org/sopt/official/feature/sopletter/onboarding/navigation/SopletterOnboardingNavigation.kt
@@ -22,34 +22,17 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.sopletter.navigation
+package org.sopt.official.feature.sopletter.onboarding.navigation
 
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
-import androidx.navigation.compose.composable
-import androidx.navigation.navigation
 import kotlinx.serialization.Serializable
-import org.sopt.official.sopletter.onboarding.SopletterOnboardingRoute
-import org.sopt.official.sopletter.onboarding.navigation.SopletterOnboarding
 
 @Serializable
-data object SopletterGraph
+data object SopletterOnboarding
 
-fun NavGraphBuilder.sopletterGraph(
-    paddingValues: PaddingValues,
-    navController: NavController,
-    navigateUp: () -> Unit,
+fun NavController.navigateToSopletterOnboarding(
+    navOptions: NavOptions? = null,
 ) {
-    navigation<SopletterGraph> (
-        startDestination = SopletterOnboarding
-    ) {
-       composable<SopletterOnboarding> {
-           SopletterOnboardingRoute (
-               paddingValues = paddingValues,
-               navigateToNickname = {}
-           )
-       }
-    }
+    navigate(SopletterOnboarding, navOptions)
 }
